Preparing file tasks
