+++ src/util/include.h
@@ -1,0 +1,10 @@
+// patrz notatke na samej górze main.cpp:
+#include <util/keys.cpp>
+#include <util/lcd.cpp>
+#include <util/misc.cpp>
+#include <util/menu.cpp>      // tworzenie, wyświetlanie i obsługa menu (LCD+keys)
+#include <util/watchdog.cpp>  // clocks, timery, cały system uzbrojenia i analizowania
+#include <util/PinsDef.h>
+
+// systemowe biblioteki
+#include <Arduino.h>
