+++ src/util/lcd.cpp
@@ -1,3 +1,4 @@
 #include <LiquidCrystal_I2C.h>
 // Konfiguracja LCD 16x2
+int backlightTime = 0; // czas podświetlenia ekranu
 LiquidCrystal_I2C lcd(0x3F,20,4);  // set the LCD address to 0x3F for a 16 chars and 2 line display
